wip: Resolve notifications MR
