+++ vector/src/main/java/im/vector/app/core/di/SingletonEntryPoint.kt
@@ -1,8 +1,17 @@
 /*
- * Copyright 2021-2024 New Vector Ltd.
+ * Copyright (c) 2021 New Vector Ltd
  *
- * SPDX-License-Identifier: AGPL-3.0-only
- * Please see LICENSE in the repository root for full details.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package im.vector.app.core.di
@@ -16,6 +25,9 @@ import im.vector.app.features.analytics.AnalyticsTracker
 import im.vector.app.features.call.webrtc.WebRtcCallManager
 import im.vector.app.features.home.AvatarRenderer
 import im.vector.app.features.navigation.Navigator
+import im.vector.app.features.notifications.utils.NotificationBuilderUtils
+import im.vector.app.features.notifications.utils.NotificationCommonUtils
+import im.vector.app.features.notifications.utils.NotificationUtils
 import im.vector.app.features.pin.PinLocker
 import im.vector.app.features.rageshake.BugReporter
 import im.vector.app.features.session.SessionListener
@@ -31,6 +43,12 @@ interface SingletonEntryPoint {
     fun sessionListener(): SessionListener
 
     fun avatarRenderer(): AvatarRenderer
+
+    fun notificationCommonUtils(): NotificationCommonUtils
+
+    fun notificationBuilderUtils(): NotificationBuilderUtils
+
+    fun notificationUtils(): NotificationUtils
 
     fun activeSessionHolder(): ActiveSessionHolder
