+++ vector/src/main/java/im/vector/app/core/services/CallRingPlayer.kt
@@ -1,8 +1,17 @@
 /*
- * Copyright 2020-2024 New Vector Ltd.
+ * Copyright (c) 2020 New Vector Ltd
  *
- * SPDX-License-Identifier: AGPL-3.0-only
- * Please see LICENSE in the repository root for full details.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package im.vector.app.core.services
@@ -21,7 +30,7 @@ import androidx.core.content.getSystemService
 import im.vector.app.R
 import im.vector.app.features.call.audio.CallAudioManager.Mode
 import im.vector.app.features.call.webrtc.WebRtcCallManager
-import im.vector.app.features.notifications.NotificationUtils
+import im.vector.app.features.notifications.utils.NotificationUtils
 import org.matrix.android.sdk.api.extensions.orFalse
 import timber.log.Timber
 
@@ -38,7 +47,7 @@ class CallRingPlayerIncoming(
 
     fun start(fromBg: Boolean) {
         val audioManager = applicationContext.getSystemService<AudioManager>()
-        val incomingCallChannel = notificationUtils.getChannelForIncomingCall(fromBg)
+        val incomingCallChannel = notificationUtils.commonUtils.getChannelForIncomingCall(fromBg)
         val ringerMode = audioManager?.ringerMode
         if (ringerMode == AudioManager.RINGER_MODE_NORMAL) {
             playRingtoneIfNeeded(incomingCallChannel)
