+++ vector/src/main/java/im/vector/app/core/services/VectorSyncAndroidService.kt
@@ -1,8 +1,17 @@
 /*
- * Copyright 2019-2024 New Vector Ltd.
+ * Copyright 2019 New Vector Ltd
  *
- * SPDX-License-Identifier: AGPL-3.0-only
- * Please see LICENSE in the repository root for full details.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package im.vector.app.core.services
 
@@ -24,7 +33,7 @@ import androidx.work.WorkerParameters
 import dagger.hilt.android.AndroidEntryPoint
 import im.vector.app.core.extensions.startForegroundCompat
 import im.vector.app.core.platform.PendingIntentCompat
-import im.vector.app.features.notifications.NotificationUtils
+import im.vector.app.features.notifications.utils.NotificationUtils
 import im.vector.app.features.settings.BackgroundSyncMode
 import im.vector.lib.core.utils.timer.Clock
 import im.vector.lib.core.utils.timer.DefaultClock
@@ -89,7 +98,7 @@ class VectorSyncAndroidService : SyncAndroidService() {
         } else {
             CommonStrings.notification_listening_for_notifications
         }
-        val notification = notificationUtils.buildForegroundServiceNotification(notificationSubtitleRes, false)
+        val notification = notificationUtils.builderUtils.buildForegroundServiceNotification(notificationSubtitleRes, false)
         startForegroundCompat(NotificationUtils.NOTIFICATION_ID_FOREGROUND_SERVICE, notification)
     }
