+++ vector/src/main/java/im/vector/app/core/utils/ExternalApplicationsUtil.kt
@@ -1,8 +1,17 @@
 /*
- * Copyright 2019-2024 New Vector Ltd.
+ * Copyright 2019 New Vector Ltd
  *
- * SPDX-License-Identifier: AGPL-3.0-only
- * Please see LICENSE in the repository root for full details.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package im.vector.app.core.utils
@@ -32,7 +41,7 @@ import androidx.core.app.ShareCompat
 import androidx.core.content.FileProvider
 import androidx.core.content.getSystemService
 import im.vector.app.R
-import im.vector.app.features.notifications.NotificationUtils
+import im.vector.app.features.notifications.utils.NotificationUtils
 import im.vector.app.features.themes.ThemeUtils
 import im.vector.lib.strings.CommonStrings
 import kotlinx.coroutines.Dispatchers
@@ -289,12 +298,12 @@ suspend fun saveMedia(
                         }
                     }
                 }
-                notificationUtils.buildDownloadFileNotification(
+                notificationUtils.builderUtils.buildDownloadFileNotification(
                         uri,
                         filename,
                         mediaMimeType ?: MimeTypes.OctetStream
                 ).let { notification ->
-                    notificationUtils.showNotificationMessage("DL", uri.hashCode(), notification)
+                    notificationUtils.commonUtils.showNotificationMessage("DL", uri.hashCode(), notification)
                 }
             }
         } else {
