+++ vector/src/main/java/im/vector/app/features/call/webrtc/ScreenCaptureAndroidService.kt
@@ -1,8 +1,17 @@
 /*
- * Copyright 2022-2024 New Vector Ltd.
+ * Copyright (c) 2022 New Vector Ltd
  *
- * SPDX-License-Identifier: AGPL-3.0-only
- * Please see LICENSE in the repository root for full details.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package im.vector.app.features.call.webrtc
@@ -13,7 +22,7 @@ import android.os.IBinder
 import dagger.hilt.android.AndroidEntryPoint
 import im.vector.app.core.extensions.startForegroundCompat
 import im.vector.app.core.services.VectorAndroidService
-import im.vector.app.features.notifications.NotificationUtils
+import im.vector.app.features.notifications.utils.NotificationUtils
 import im.vector.lib.core.utils.timer.Clock
 import javax.inject.Inject
 
@@ -32,7 +41,7 @@ class ScreenCaptureAndroidService : VectorAndroidService() {
 
     private fun showStickyNotification() {
         val notificationId = clock.epochMillis().toInt()
-        val notification = notificationUtils.buildScreenSharingNotification()
+        val notification = notificationUtils.builderUtils.buildScreenSharingNotification()
         startForegroundCompat(notificationId, notification)
     }
