+++ vector/src/main/java/im/vector/app/features/settings/troubleshoot/TestNotification.kt
@@ -1,15 +1,24 @@
 /*
- * Copyright 2018-2024 New Vector Ltd.
+ * Copyright 2018 New Vector Ltd
  *
- * SPDX-License-Identifier: AGPL-3.0-only
- * Please see LICENSE in the repository root for full details.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package im.vector.app.features.settings.troubleshoot
 
 import android.content.Context
 import im.vector.app.core.resources.StringProvider
 import im.vector.app.core.utils.startNotificationSettingsIntent
-import im.vector.app.features.notifications.NotificationUtils
+import im.vector.app.features.notifications.utils.NotificationUtils
 import im.vector.lib.strings.CommonStrings
 import javax.inject.Inject
 
@@ -25,7 +34,7 @@ class TestNotification @Inject constructor(
 
     override fun perform(testParameters: TestParameters) {
         // Display the notification right now
-        notificationUtils.displayDiagnosticNotification()
+        notificationUtils.commonUtils.displayDiagnosticNotification()
         description = stringProvider.getString(CommonStrings.settings_troubleshoot_test_notification_notice)
 
         quickFix = object : TroubleshootQuickFix(CommonStrings.open_settings) {
