+++ vector/src/test/java/im/vector/app/test/fakes/FakeNotificationUtils.kt
@@ -9,8 +9,8 @@ package im.vector.app.test.fakes
 
 import android.app.Notification
 import im.vector.app.features.notifications.InviteNotifiableEvent
-import im.vector.app.features.notifications.NotificationUtils
 import im.vector.app.features.notifications.SimpleNotifiableEvent
+import im.vector.app.features.notifications.utils.NotificationUtils
 import io.mockk.every
 import io.mockk.mockk
 
@@ -20,13 +20,13 @@ class FakeNotificationUtils {
 
     fun givenBuildRoomInvitationNotificationFor(event: InviteNotifiableEvent, myUserId: String): Notification {
         val mockNotification = mockk<Notification>()
-        every { instance.buildRoomInvitationNotification(event, myUserId) } returns mockNotification
+        every { instance.builderUtils.buildRoomInvitationNotification(event, myUserId) } returns mockNotification
         return mockNotification
     }
 
     fun givenBuildSimpleInvitationNotificationFor(event: SimpleNotifiableEvent, myUserId: String): Notification {
         val mockNotification = mockk<Notification>()
-        every { instance.buildSimpleEventNotification(event, myUserId) } returns mockNotification
+        every { instance.builderUtils.buildSimpleEventNotification(event) } returns mockNotification
         return mockNotification
     }
 }
